watch mode: fix watch/unwatch on deps chagne
We keep a reference to the old addresses and we compare the newAddresses after
a watch file is changed. If we edited the dependencies then there may be new
dependencies added that we should watch or removed dependencies that we should
stop watching. We apply _.difference twice to update fs notifiers.